updated print styles, purely in browser solution

+++ src/components/content-section/content-section.ts
@@ -91,7 +91,9 @@ export class ContentSection implements OnInit {
 	}
 
 	removeSection() {
-		this.contentService.removeContentSection(this.id());
+		if (window.confirm("Are you sure you want to delete section?")) {
+			this.contentService.removeContentSection(this.id());
+		}
 	}
 
 	sectionSelected() {
